refactor(MessageService): clean code and extract business logic to message_controller for testability
