+++ Backend/MessageService/Server/server.cpp
@@ -6,14 +6,23 @@
 Server::Server(int port, MessageManager* message_manager,
                RabbitMQClient* mq_client)
     : port_(port) {
-  LOG_INFO("Start: Controller created");
   controller_ = std::make_unique<Controller>(app_, mq_client, message_manager);
-  LOG_INFO("Controller created");
   handleRoutes();
-  LOG_INFO("Routes handled");
 }
 
-void Server::handleRoutes() { controller_->handleRoutes(); }
+void Server::handleRoutes() {
+  handleGetMessagesFromChat();
+}
+
+void Server::handleGetMessagesFromChat() {
+  CROW_ROUTE(app_, "/messages/<int>")
+  .methods(crow::HTTPMethod::GET)([&](const crow::request& req, crow::response& res, int chat_id) {
+    PROFILE_SCOPE("/message/<int>");
+    controller_->getMessagesFromChat(req, chat_id, res);
+    LOG_INFO("Response code: {} | Body: {}", res.code, res.body);
+    });
+}
+
 
 void Server::run() {
   spdlog::info("[Message server is started on port '{}'", port_);
