+++ Frontend/src/NetworkManagers/MessageManager/messagemanager.cpp
@@ -28,7 +28,7 @@ QFuture<QList<Message>> MessageManager::getChatMessages(const QString& current_t
            before_id);
   QUrlQuery query;
   query.addQueryItem("limit", QString::number(limit));
-  query.addQueryItem("beforeId", QString::number(before_id));
+  query.addQueryItem("before_id", QString::number(before_id));
   endpoint.setQuery(query);
   auto request = getRequestWithToken(endpoint, current_token);
   auto* reply = network_manager_->get(request); //TODO(roma): make function getReplyGetChatMessages();
